+++ doc/Changelog.md
@@ -13,6 +13,7 @@
   - Rename `InsulationType` into `InsulatorType`.
   - Allow the letter "U" for "Underground" line type (only "S" for "Souterrain" in French was accepted). The same
     with the letter "O" for "Overhead" line type (only "A" for "Aérien" in French was accepted).
+  - Remove the field `"model"` from the JSON serialization of `LineParameters`.
 - [PR120](https://github.com/RoseauTechnologies/Roseau_Load_Flow/pull/120) Fix phases of flexible power results.
 - [PR119](https://github.com/RoseauTechnologies/Roseau_Load_Flow/pull/119) Add explicit error message for singular jacobian.
 - [PR118](https://github.com/RoseauTechnologies/Roseau_Load_Flow/pull/118) [GH95](https://github.com/RoseauTechnologies/Roseau_Load_Flow/issues/95)